+++ tests/qml/Common/Theme.qml
@@ -1,0 +1,14 @@
+pragma Singleton
+import QtQuick
+QtObject {
+    property int spacingS: 8
+    property int spacingM: 12
+    property int fontSizeLarge: 18
+    property int fontSizeMedium: 14
+    property int fontSizeSmall: 12
+    property int cornerRadius: 8
+    property color surfaceText: "white"
+    property color surfaceVariantText: "gray"
+    property color surfaceContainerHigh: "black"
+    property color error: "red"
+}